Revert "Ngmix cleanup"
