+++ example/cfis/config_tile_Ng_template.ini
@@ -71,9 +71,6 @@ LOG_WCS = $SP_RUN/output/log_exp_headers.sqlite
 # Magnitude zero-point
 MAG_ZP = 30.0
 
-# Pixel scale in arcsec
-PIXEL_SCALE = 0.186
-
 ID_OBJ_MIN = X
 ID_OBJ_MAX = X
