+++ shapepipe/modules/merge_sep_cats_package/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""MERGE SEP CATS PACKAGE
+"""MERGE SEP CAT PACKAGE
 
 This package contains the module for ``merge_sep_cat``.
