--- shapepipe/modules/ngmix_package/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-""" NGMIX PACKAGE
-
-This package contains the code for ``ngmix``.
-
-:Author: Axel Guinot
-
-"""
-
-__all__ = ['ngmix']
